telegram_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
nest-asyncio==1.5.8